TPs/TP5/Pilar2/mineros/mineroCPU/requirements.txt
requests
python-dateutil
pydantic==2.7.1
websocket-client